# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Kubernetes Cluster Info Collector API v2.0
  description: |
    A comprehensive REST API for querying Kubernetes cluster information collected by the Cluster Info Collector v2.0.
    
    ## Features
    - **9 Resource Types**: Deployments, Pods, Nodes, Services, Ingresses, ConfigMaps, Secrets, PersistentVolumes, PersistentVolumeClaims
    - **Dual Architecture**: Legacy direct-write and Kafka-based streaming
    - **Real-time Streaming**: WebSocket API for live cluster data
    - **Historical Data**: Time-series analysis with configurable retention
    - **Monitoring**: Prometheus metrics and health endpoints
    - **Data Retention**: Automated cleanup with configurable policies
    
    ## Architecture Modes
    - **Legacy Mode**: Direct database writes (simple deployment)
    - **Kafka Mode**: Producer-consumer pattern with horizontal scaling
    
    **Base URL**: `http://localhost:8081/api/v1`
    
    **Authentication**: None required (read-only API)
    
    **Rate Limiting**: Not implemented
    
    **CORS**: Enabled for all origins
    
    **WebSocket**: Available at `/api/v1/ws` for real-time data streaming
  version: 2.0.0
  contact:
    name: Kubernetes Cluster Info Collector
    url: https://github.com/srikarm/k8s-cluster-info-collector
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Local development server
...
    description: Production server
...
paths:
  /snapshots:
    get:
      tags:
        - snapshots
      summary: List cluster snapshots
      description: Retrieve a list of cluster snapshots with resource counts
      parameters:
        - name: limit
          in: query
          description: Maximum number of snapshots to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
      responses:
        '200':
          description: List of snapshots
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshots:
                    type: array
                    items:
                      $ref: '#/components/schemas/SnapshotSummary'
                  count:
                    type: integer
                    description: Number of snapshots returned
              example:
                snapshots:
                  - id: 123
                    timestamp: "2025-01-15T10:30:00Z"
                    deployments: 15
                    pods: 45
                    nodes: 3
                    services: 20
                    ingresses: 5
                    configmaps: 25
                    secrets: 30
                    persistent_volumes: 10
                    persistent_volume_claims: 8
                count: 1
        '500':
          $ref: '#/components/responses/InternalServerError'
  /snapshots/{id}:
    get:
      tags:
        - snapshots
      summary: Get specific snapshot
      description: Retrieve detailed information for a specific cluster snapshot
      parameters:
        - name: id
          in: path
          required: true
          description: Snapshot ID
          schema:
            type: integer
      responses:
        '200':
          description: Detailed snapshot information
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  timestamp:
                    type: string
                    format: date-time
                  cluster_info:
                    $ref: '#/components/schemas/ClusterInfo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /snapshots/latest:
    get:
      tags:
        - snapshots
      summary: Get latest snapshot
      description: Redirect to the most recent cluster snapshot
      responses:
        '302':
          description: Redirects to /snapshots/{id} for the latest snapshot
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /deployments:
    get:
      tags:
        - resources
      summary: List deployments
      description: Get all deployments from the latest snapshot
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/NamespaceParam'
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeploymentSummary'
                  count:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /pods:
    get:
      tags:
        - resources
      summary: List pods
      description: Get all pods from the latest snapshot
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/NamespaceParam'
      responses:
        '200':
          description: List of pods
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PodSummary'
                  count:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /nodes:
    get:
      tags:
        - resources
      summary: List nodes
      description: Get all nodes from the latest snapshot
      parameters:
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of nodes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NodeSummary'
                  count:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /services:
    get:
      tags:
        - resources
      summary: List services
      description: Get all services from the latest snapshot
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/NamespaceParam'
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServiceSummary'
                  count:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /ingresses:
    get:
      tags:
        - resources
      summary: List ingresses
      description: Get all ingresses from the latest snapshot
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/NamespaceParam'
      responses:
        '200':
          description: List of ingresses
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/IngressSummary'
                  count:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /configmaps:
    get:
      tags:
        - resources
      summary: List ConfigMaps
      description: Get all ConfigMaps from the latest snapshot
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/NamespaceParam'
      responses:
        '200':
          description: List of ConfigMaps
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConfigMapSummary'
                  count:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /secrets:
    get:
      tags:
        - resources
      summary: List Secrets
      description: Get all Secrets from the latest snapshot (data keys only, not values)
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/NamespaceParam'
      responses:
        '200':
          description: List of Secrets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SecretSummary'
                  count:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /persistent-volumes:
    get:
      tags:
        - resources
      summary: List PersistentVolumes
      description: Get all PersistentVolumes from the latest snapshot
      parameters:
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of PersistentVolumes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PersistentVolumeSummary'
                  count:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /persistent-volume-claims:
    get:
      tags:
        - resources
      summary: List PersistentVolumeClaims
      description: Get all PersistentVolumeClaims from the latest snapshot
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/NamespaceParam'
      responses:
        '200':
          description: List of PersistentVolumeClaims
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PersistentVolumeClaimSummary'
                  count:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /stats:
    get:
      tags:
        - statistics
      summary: Get general statistics
      description: Retrieve general cluster statistics including total snapshots and latest snapshot resource counts
      responses:
        '200':
          description: General statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_snapshots:
                    type: integer
                    description: Total number of snapshots stored
                  latest_snapshot:
                    type: object
                    description: Resource counts from the latest snapshot
                    properties:
                      deployments:
                        type: integer
                      pods:
                        type: integer
                      nodes:
                        type: integer
                      services:
                        type: integer
                      ingresses:
                        type: integer
                      configmaps:
                        type: integer
                      secrets:
                        type: integer
                      persistent_volumes:
                        type: integer
                      persistent_volume_claims:
                        type: integer
              example:
                total_snapshots: 150
                latest_snapshot:
                  deployments: 15
                  pods: 45
                  nodes: 3
                  services: 20
                  ingresses: 5
                  configmaps: 25
                  secrets: 30
                  persistent_volumes: 10
                  persistent_volume_claims: 8
        '500':
          $ref: '#/components/responses/InternalServerError'
  /stats/retention:
    get:
      tags:
        - statistics
      summary: Get retention statistics
      description: Retrieve data retention statistics including database size and snapshot age range
      responses:
        '200':
          description: Retention statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  database_size:
                    type: string
                    description: Human-readable database size
                  oldest_snapshot:
                    type: string
                    format: date-time
                    description: Timestamp of the oldest snapshot
                  newest_snapshot:
                    type: string
                    format: date-time
                    description: Timestamp of the newest snapshot
                  retention_span:
                    type: string
                    description: Time span between oldest and newest snapshots
              example:
                database_size: "256 MB"
                oldest_snapshot: "2025-01-01T00:00:00Z"
                newest_snapshot: "2025-01-15T12:00:00Z"
                retention_span: "14d12h"
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /ws:
    get:
      tags:
        - streaming
      summary: WebSocket connection
      description: |
        Establish a WebSocket connection for real-time cluster data streaming.
        
        **Protocol**: WebSocket (ws://)
        
        **Message Types**:
        - `cluster_update`: Full cluster data update
        - `metrics_update`: Metrics data update
        - `alert`: Alert notifications
        
        **Connection**: Upgrade HTTP connection to WebSocket
      responses:
        '101':
          description: WebSocket connection established
        '503':
          description: WebSocket streaming not enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of resources to return
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    NamespaceParam:
      name: namespace
      in: query
      description: Filter resources by namespace
      required: false
      schema:
        type: string
  schemas:
    SnapshotSummary:
      type: object
      properties:
        id:
          type: integer
          description: Unique snapshot identifier
        timestamp:
          type: string
          format: date-time
          description: When the snapshot was taken
        deployments:
          type: integer
          description: Number of deployments in this snapshot
        pods:
          type: integer
          description: Number of pods in this snapshot
        nodes:
          type: integer
          description: Number of nodes in this snapshot
        services:
          type: integer
          description: Number of services in this snapshot
        ingresses:
          type: integer
          description: Number of ingresses in this snapshot
        configmaps:
          type: integer
          description: Number of ConfigMaps in this snapshot
        secrets:
          type: integer
          description: Number of Secrets in this snapshot
        persistent_volumes:
          type: integer
          description: Number of PersistentVolumes in this snapshot
        persistent_volume_claims:
          type: integer
          description: Number of PersistentVolumeClaims in this snapshot
    ClusterInfo:
      type: object
      description: Complete cluster state information
      properties:
        timestamp:
          type: string
          format: date-time
        deployments:
          type: array
          items:
            $ref: '#/components/schemas/DeploymentInfo'
        pods:
          type: array
          items:
            $ref: '#/components/schemas/PodInfo'
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeInfo'
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceInfo'
        ingresses:
          type: array
          items:
            $ref: '#/components/schemas/IngressInfo'
        configmaps:
          type: array
          items:
            $ref: '#/components/schemas/ConfigMapInfo'
        secrets:
          type: array
          items:
            $ref: '#/components/schemas/SecretInfo'
        persistent_volumes:
          type: array
          items:
            $ref: '#/components/schemas/PersistentVolumeInfo'
        persistent_volume_claims:
          type: array
          items:
            $ref: '#/components/schemas/PersistentVolumeClaimInfo'
...
    DeploymentSummary:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        replicas:
          type: integer
        ready_replicas:
          type: integer
        created_time:
          type: string
          format: date-time
...
    PodSummary:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        phase:
          type: string
        node_name:
          type: string
        restart_count:
          type: integer
        created_time:
          type: string
          format: date-time
...
    NodeSummary:
      type: object
      properties:
        name:
          type: string
        ready:
          type: boolean
        cpu_capacity:
          type: string
        memory_capacity:
          type: string
        created_time:
          type: string
          format: date-time
...
    ServiceSummary:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        type:
          type: string
        cluster_ip:
          type: string
        created_time:
          type: string
          format: date-time
...
    IngressSummary:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        hosts:
          type: array
          items:
            type: string
        created_time:
          type: string
          format: date-time
...
    ConfigMapSummary:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        data_keys:
          type: array
          items:
            type: string
        created_time:
          type: string
          format: date-time
...
    SecretSummary:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        type:
          type: string
        data_keys:
          type: array
          items:
            type: string
        created_time:
          type: string
          format: date-time
...
    PersistentVolumeSummary:
      type: object
      properties:
        name:
          type: string
        capacity:
          type: string
        access_modes:
          type: array
          items:
            type: string
        status:
          type: string
        storage_class:
          type: string
        created_time:
          type: string
          format: date-time
...
    PersistentVolumeClaimSummary:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        requested_size:
          type: string
        access_modes:
          type: array
          items:
            type: string
        status:
          type: string
        created_time:
          type: string
          format: date-time
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid snapshot ID"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Snapshot not found"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Failed to fetch data"
...